director/requirements.txt
Django>2.0,<2.1
django-configurations==2.0
django-extensions==2.0.7
django-polymorphic==2.0.2
djangorestframework==3.8.2
dj_database_url==0.5.0
requests==2.19.1
sh==1.12.14
